MES-3950: Refactor of test data selector + Added BE selector
