+++ src/pages/test-report/cat-b/components/manoeuvre-competency/manoeuvre-competency.ts
@@ -10,7 +10,7 @@ import {
 import { getCurrentTest } from '../../../../../modules/tests/tests.selector';
 import { getTestData } from '../../../../../modules/tests/test-data/test-data.reducer';
 import { getTests } from '../../../../../modules/tests/tests.reducer';
-import { getManoeuvres } from '../../../../../modules/tests/test-data/test-data.selector';
+import { getManoeuvres } from '../../../../../modules/tests/test-data/cat-b/test-data.cat-b.selector';
 import { getTestReportState } from '../../../test-report.reducer';
 import { isRemoveFaultMode, isSeriousMode, isDangerousMode } from '../../../test-report.selector';
 import { manoeuvreCompetencyLabels } from '../../../../../shared/constants/competencies/catb-manoeuvres';
