+++ src/pages/test-report/cat-b/components/manoeuvres-popover/manoeuvres-popover.ts
@@ -4,7 +4,7 @@ import { Observable } from 'rxjs/Observable';
 import { Component } from '@angular/core';
 import { getCurrentTest } from '../../../../../modules/tests/tests.selector';
 import { getTestData } from '../../../../../modules/tests/test-data/test-data.reducer';
-import { getManoeuvres } from '../../../../../modules/tests/test-data/test-data.selector';
+import { getManoeuvres } from '../../../../../modules/tests/test-data/cat-b/test-data.cat-b.selector';
 import { getTests } from '../../../../../modules/tests/tests.reducer';
 import { StoreModel } from '../../../../../shared/models/store.model';
 import { RecordManoeuvresSelection } from '../../../../../modules/tests/test-data/manoeuvres/manoeuvres.actions';
