+++ src/pages/waiting-room-to-car/cat-b/waiting-room-to-car.cat-b.page.ts
@@ -59,15 +59,17 @@ import {
   EyesightTestFailed,
 } from '../../../modules/tests/test-data/eyesight-test/eyesight-test.actions';
 import {
+  hasEyesightTestBeenCompleted,
+} from '../../../modules/tests/test-data/test-data.selector';
+import {
   isTellMeQuestionSelected,
   isTellMeQuestionDrivingFault,
   isTellMeQuestionCorrect,
   tellMeQuestionOutcome,
   getVehicleChecks,
   getTellMeQuestion,
-  hasEyesightTestBeenCompleted,
   hasEyesightTestGotSeriousFault,
-} from '../../../modules/tests/test-data/test-data.selector';
+} from '../../../modules/tests/test-data/cat-b/test-data.cat-b.selector';
 import { getTestData } from '../../../modules/tests/test-data/test-data.reducer';
 import { PersistTests } from '../../../modules/tests/tests.actions';
 import { CAT_B } from '../../page-names.constants';
